+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -4,6 +4,8 @@ Release History
 ===============
 (unreleased)
 +++++++++++++++++++
+* availability-set: expose fault domain count on convert
+* vm: expose 'az vm list-skus' command
 * vm/vmss: support to assign identity w/o creating role assignments
 * vm: apply storage sku on attaching data disks